params.build to specific directory passed into constructor (dir)
In the constructor for graphBuilder.fordirectory, the specific directory is passed in as dir from OTPMain.java. I changed a reference to the old params.build, which was a File (it is now initiated as a List<File> in CommandLineParameters.java to support multiple directories) to the specific directory, dir, which is passed into the forDirectory constructor with params.build.get(i) in a for loop in OTPMain.java.